+++ apps/example-app/CHANGELOG.md
@@ -5,4 +5,11 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ianduvall/resize-observer-store@0.0.7
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
   - @ianduvall/resize-observer-store@0.0.6