+++ apps/react-app/CHANGELOG.md
@@ -5,4 +5,11 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ianduvall/react-resize-observer-store@0.0.5
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
   - @ianduvall/react-resize-observer-store@0.0.4