+++ packages/resize-observer-store/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ianduvall/resize-observer-store
 
+## 0.0.7
+
+### Patch Changes
+
+- Refactor resize-observer-store to export functions instead of store instance
+- Updated dependencies
+  - @ianduvall/subscribable@0.0.1
+  - @ianduvall/lazily@0.0.1
+
 ## 0.0.6
 
 ### Patch Changes